refactor(desktop): split inference setup options
