+++ packages/gents-desktop-fleet/src/inference/useInferenceSetup.ts
@@ -1,12 +1,8 @@
 import { useCallback, useEffect, useState } from "react";
 
 import type {
-  BackendSaveRequest,
-  BehaviorSaveRequest,
   CodexLoginResult,
-  DeploymentView,
   GrokLoginResult,
-  InferenceProbeResult,
 } from "@source-inc/gents-desktop-client";
 import {
   CODEX_DEFAULT_MODEL,
@@ -30,25 +26,9 @@ import {
   type PersistBackendOptions,
 } from "./persistBackend.js";
 import { resolveTargets } from "./resolveTargets.js";
+import type { InferenceSetupOptions } from "./types.js";
 
-export type InferenceSetupOptions = {
-  deployment: DeploymentView;
-  onClose: () => void;
-  onSaveBackendConfig: (request: BackendSaveRequest) => Promise<unknown>;
-  onSaveBehaviorConfig: (request: BehaviorSaveRequest) => Promise<unknown>;
-  onProbeInferenceEndpoint: (endpoint: string) => Promise<InferenceProbeResult>;
-  onCodexLogin: (agentDid: string) => Promise<CodexLoginResult>;
-  /** Abort a ChatGPT sign-in whose browser was closed, so it does not hang. */
-  onCancelCodexLogin?: () => Promise<unknown>;
-  onCodexLoginUrl?: (
-    onUrl: (url: string | null) => void,
-  ) => Promise<() => void>;
-  onGrokLogin?: (agentDid: string) => Promise<GrokLoginResult>;
-  onCancelGrokLogin?: () => Promise<unknown>;
-  onGrokLoginUrl?: (
-    onUrl: (url: string | null) => void,
-  ) => Promise<() => void>;
-};
+export type { InferenceSetupOptions } from "./types.js";
 
 export function useInferenceSetup({
   deployment,
